refactor: move pkg to `gitea_cc_plugin`
